build: add credential provider logic back
